math/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.0.0",
    "pytest>=9.0.1",
    "uvicorn>=0.27.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]
